# Add 2K software prefetch to improve BITCOUNT performance