## Summary

Adds a software prefetch with a 2K stride to the scalar popcount loop in redisPopcount().
This reduces L3-bound stalls and improves ops/sec by ~25% on Intel Sapphire Rapids,
with no negative impact on L1/L2 usage or cache pollution (confirmed with HW counters). 

**Note:** The 2K stride was the best starting from 128,256,512,1024,2048,4096. 
4K gave the same outcome so it's best to avoid larger strides without reason.


## Performance gains on client side

- +25.6% throughput gain: +6,789 ops/sec with prefetching
- -20.1% latency reduction: Avg latency dropped from 7.53 ms to 6.01 ms

Configuration | Avg Ops/sec | Avg Latency (ms) | p50 Latency | p99 Latency | p99.9 Latency | Avg KB/sec
-- | -- | -- | -- | -- | -- | --
Baseline b7c6755b1b44da36866c0376da0a000873177270 (no prefetch) | 26,550.91 | 7.531 | 7.839 | 12.863 | 14.527 | 2,881.61
With 2K Prefetch | 33,339.50 | 6.014 | 6.239 | 10.239 | 11.583 | 3,618.39




## Performance gains on HW

- L3 stalls decreased, confirmed via Topdown metrics

Metric | Baseline (%) | Prefetch (%) | Notes
-- | -- | -- | --
Backend_Bound | 55.5 | 43.9 | 🟢 Reduced overall backend stalls — shows better pipeline throughput
└── Memory_Bound | 23.0 | 11.4 | 🟢 Major drop in memory stalls — prefetching mitigated memory latency
├── L1_Bound | 1.6 | 3.1 | 🔺 Slightly increased — prefetch might hit L1, but not harmful
├── L2_Bound | 0.3 | 2.1 | 🔺 Higher, likely due to deeper prefetch coverage
├── L3_Bound | 20.5 | 6.1 | 🟢 Huge reduction — prefetch hides L3 latency well
└── DRAM_Bound | 0.0 | 0.0 | No significant DRAM pressure in either case
Core_Bound | 32.4 | 32.5 | ⚪️ No change — compute capacity utilization unchanged
└── Ports_Utilization | 23.9 | 21.8 | 🟢 Slight improvement in execution port efficiency
Frontend_Bound | 3.1 | 2.7 | 🟢 Minor improvement — better i-cache behavior
Retiring | 41.0 | 52.9 | 🟢 More uops completed — clearer sign of improved efficiency

## Replicate the results

### Benchmark

preload
```
 memtier_benchmark --hide-histogram --command "SETBIT users __key__ 1" --key-maximum 100000000 --key-minimum 1 -n 10000 --key-prefix "" --command-key-pattern R --distinct-client-seed --pipeline 10
```

benchmark
```
memtier_benchmark --command "BITCOUNT users __key__ -1" --key-maximum 100000000 --key-minimum 1 --key-prefix "" --command-key-pattern R --distinct-client-seed --test-time 60  --hide-histogram -x 3
```

### HW counters

you can replicate the hw collection with the following command on the server side while benchmarking (you need a metal VM of physical HW access)
```
 sudo ./toplev.py --pid $(pgrep redis-server) --level 3   --nodes '+Backend_Bound.Memory_Bound,+Backend_Bound.Memory_Bound.L1_Bound,+Backend_Bound.Memory_Bound.L2_Bound,+Backend_Bound.Memory_Bound.L3_Bound'  --verbose -- sleep 30
```