@fcostaoliveira 2025-06-04
### CE Performance Automation : step 1 of 2 (build) STARTING...

This comment was automatically generated given a benchmark was triggered.
Started building at 2025-06-04 20:31:54.078680
You can check each build/benchmark progress in grafana:
   - git hash: b21fca9fb4828236a645f2cda452bc3227468665
   - git branch: filipecosta90:prefetch.popcount
   - commit date and time: n/a
   - commit summary: n/a
   - test filters:
       - command priority lower limit: 0
       - command priority upper limit: 10000
       - test name regex: .*
       - command group regex: .*